#1719 Moving of oos-branding classes to a package (org.datacleaner.branding).
