+++ oss-branding/src/main/java/org/datacleaner/branding/ComponentInstanceResourceBundle.java
@@ -18,10 +18,12 @@
  * Boston, MA  02110-1301  USA
  */
 
+package org.datacleaner.branding;
+
 import java.io.IOException;
 
-public class MonitorResourceBundle extends ImageLoadingPropertyResourceBundle {
-    public MonitorResourceBundle() throws IOException {
-        super(MonitorResourceBundle.class.getClassLoader().getResource("monitor-images.properties"));
+public class ComponentInstanceResourceBundle extends ImageLoadingPropertyResourceBundle {
+    public ComponentInstanceResourceBundle() throws IOException {
+        super(ComponentInstanceResourceBundle.class.getClassLoader().getResource("component-images.properties"));
     }
 }
